requirements.txt
fastapi
uvicorn
jinja2
torch
tensorflow
pandas
numpy
mysql-connector-python
bcrypt
python-jose[cryptography]
python-multipart
python-dotenv